// WebInventoryPortal/inventory/src/app/models/inventoryorderitem.ts
export class inventoryorderitem{
    orderID: number ;
    inventoryID : number ;
    inventoryItemID   : number ;
    email :     string ;
    project:    string ;
    instDate:    Date ;
    destBuilding:     string ;
    destFloor:     string ;
    destRoom:    string ;
    qty     : number ;
    category:    string ;
    itemCode :    string ;
    description:     string ;
    building:   string ;
    floor:    string ;
    room:     string ;
    imageBase64 :    string ;
    imgBase64Num: number;
    instDateStr: string;
}